pyproject.toml
[project]
name = "cluster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "psycopg2-binary>=2.9.9",
    "pyyaml>=6.0.0",
]
